+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+
 ## 1.0.0
 
 Initial public release.
@@ -19,3 +20,26 @@ Initial public release.
 ## 1.0.2
 
 - Update for Certification standards
+
+
+
+## [1.1.0] - 2026-07-17
+
+### Added
+- Homey Flow Condition: Glucose is above
+- Homey Flow Condition: Glucose is below
+- Homey Flow Condition: Glucose is between
+- Homey Flow Condition: Trend is
+- Homey Flow Condition: Reservoir is below
+- Homey Flow Condition: Pump battery is below
+- Homey Flow Condition: CGM battery is below
+- Homey Flow Condition: Nightscout is available
+
+### Improved
+- Automatic mmol/L and mg/dL support in Flow Conditions.
+- Better handling of temporary Nightscout connection failures.
+- Improved application stability during startup.
+
+### Removed
+- Removed obsolete IOB capability definitions.
+- Removed obsolete COB capability definitions.